% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_comma}
\alias{add_comma}
\title{Add Comma Separators to Numbers}
\usage{
add_comma(x, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{...}{Additional arguments passed to format()}
}
\value{
Character vector with formatted numbers
}
\description{
Formats numbers with comma separators for thousands.
}
\examples{
add_comma(1234567)
add_comma(c(1000, 2000000, 500))
}
