% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_unnest.R
\name{col_unnest}
\alias{col_unnest}
\title{Unnest a column containing vectors}
\source{
https://stackoverflow.com/questions/40420597/r-flatten-nested-data-table/40420690#40420690
}
\usage{
col_unnest(x, col)
}
\arguments{
\item{x}{tibble. database with a column to unnest.}

\item{col}{character. Name of the column to unnest.}
}
\value{
A dataframe where all variables are discrete.
}
\description{
Faster implementation of unnest applicable only when the nested column is a vector (i.e. not a table or dataframe).
}
\author{
Nicolas Mangin
}
