% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-detection.R
\name{format-detection}
\alias{format-detection}
\title{Output Format Detection}
\value{
A character string indicating the output format
}
\description{
These functions detect the current R Markdown output format and return
the appropriate format string for various table packages.
}
