% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-detection.R
\name{get_kable_format}
\alias{get_kable_format}
\title{Detect format for kable}
\usage{
get_kable_format()
}
\description{
Returns "latex", "html", or "markdown" depending on knitr output format.
}
\examples{
get_kable_format()
}
