% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-labels.R
\name{get_label_mappings}
\alias{get_label_mappings}
\title{Get Current Label Mappings}
\usage{
get_label_mappings()
}
\value{
A named character vector of mappings, or NULL if none set.
}
\description{
Retrieve the currently registered custom label mappings.
}
\examples{
get_label_mappings()
}
