% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-detection.R
\name{get_star_format}
\alias{get_star_format}
\title{Detect format for stargazer}
\usage{
get_star_format()
}
\description{
Returns "latex", "html", or "text" depending on knitr output format.
}
\examples{
get_star_format()
}
