% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-detection.R
\name{get_xtable_format}
\alias{get_xtable_format}
\title{Detect format for xtable}
\usage{
get_xtable_format()
}
\description{
Returns "latex" or "html" depending on knitr output format.
}
\examples{
get_xtable_format()
}
