% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmer-stargazer.R
\name{glmer-stargazer}
\alias{glmer-stargazer}
\title{Stargazer Support for glmer Models}
\description{
Functions to create stargazer-compatible model objects from glmer models.
Two approaches are provided:
\enumerate{
\item glmer_to_stargazer() - creates a "trojan" glm object (simpler, some limitations)
\item star_glmer() - wrapper that uses stargazer's coef/se override (recommended)
}
}
