% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmer-stargazer.R
\name{glmer_list_to_stargazer}
\alias{glmer_list_to_stargazer}
\title{Convert Multiple glmer Models for Stargazer}
\usage{
glmer_list_to_stargazer(models, exponentiate = FALSE)
}
\arguments{
\item{models}{A list of glmer models}

\item{exponentiate}{Logical. If TRUE, exponentiates coefficients.}
}
\value{
A list of trojan glm objects
}
\description{
Convenience function to convert a list of glmer models.
}
\examples{
\dontrun{
models <- list(model1, model2, model3)
trojans <- glmer_list_to_stargazer(models, exponentiate = TRUE)
stargazer(trojans)
}
}
