% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not In Operator}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{Vector of values to check}

\item{table}{Vector of values to check against}
}
\value{
Logical vector
}
\description{
Negation of the \\%in\\% operator. Returns TRUE for elements not in the set.
}
\examples{
c(1, 2, 3) \%nin\% c(2, 4)
# Returns: TRUE FALSE TRUE
}
