% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique.R
\name{make_unique}
\alias{make_unique}
\title{Force unique observation}
\usage{
make_unique(x)
}
\arguments{
\item{x}{Dataframe. All entries for one observation.}
}
\value{
Dataframe with only one row for the observation.
}
\description{
Remove duplicated entries. If there are several discrepant pieces of information about the observation, take the most frequent unique value, or the mean, or a random value of the duplicates to keep only one observation.
}
\author{
Nicolas Mangin
}
