% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.r
\name{mean_sd}
\alias{mean_sd}
\title{output mean and standard deviation in a nice format}
\usage{
mean_sd(dat, mean.sd.pars, dp, p1, p2)
}
\arguments{
\item{dat}{your cohort data file, of \code{data.frame} structure. For more details please refer the iris dataset.
\tabular{rrrrl}{
  \strong{Sepal.Length} \tab \strong{Sepal.Width} \tab \strong{Petal.Length} \tab \strong{Petal.Width} \tab \strong{Species} \cr
  5.1 \tab 3.5 \tab 1.4 \tab 0.2 \tab setosa\cr
  4.9 \tab 3.0 \tab 1.4 \tab 0.2 \tab setosa\cr
  4.7 \tab 3.2 \tab 1.3 \tab 0.2 \tab setosa\cr
  4.6 \tab 3.1 \tab 1.5 \tab 0.2 \tab setosa\cr
  5.0 \tab 3.6 \tab 1.4 \tab 0.2 \tab setosa\cr
  5.4 \tab 3.9 \tab 1.7 \tab 0.4 \tab setosa
}}

\item{mean.sd.pars}{a \code{character} vector which stores all parameter name that need mean sd calculation.}

\item{dp}{a \code{numeric} variable, number of decimal places to display.}

\item{p1}{a \code{string} variable, the separator appeared between mean and sd. i.e. \code{<mean><p1><sd><p2>}}

\item{p2}{a \code{string} variable, the separator appeared after sd. i.e. \code{<mean><p1><sd><p2>}}
}
\description{
This function gives a table of listing parameters' mean and sd
}
\examples{
\dontrun{
# d<-dat
# p<-get.stat.par(par,par$Parameters))
# mean.sd.p<-as.vector(p[["Mean.Sd."]])
# mean.sd.dp<-2
# mean.sd.p1<-"("
# mean.sd.p2<-")"
# mean_sd(d,mean.sd.p,mean.sd.dp,mean.sd.p1,mean.sd.p2)}

}
