% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_iqr.r
\name{median_iqr}
\alias{median_iqr}
\title{output median and iqr in a nice format}
\usage{
median_iqr(dat, mean.sd.pars, dp, p1, p2, p3)
}
\arguments{
\item{dat}{your cohort data file, of \code{data.frame} structure. For more details please refer the iris dataset.
\tabular{rrrrl}{
  \strong{Sepal.Length} \tab \strong{Sepal.Width} \tab \strong{Petal.Length} \tab \strong{Petal.Width} \tab \strong{Species} \cr
  5.1 \tab 3.5 \tab 1.4 \tab 0.2 \tab setosa\cr
  4.9 \tab 3.0 \tab 1.4 \tab 0.2 \tab setosa\cr
  4.7 \tab 3.2 \tab 1.3 \tab 0.2 \tab setosa\cr
  4.6 \tab 3.1 \tab 1.5 \tab 0.2 \tab setosa\cr
  5.0 \tab 3.6 \tab 1.4 \tab 0.2 \tab setosa\cr
  5.4 \tab 3.9 \tab 1.7 \tab 0.4 \tab setosa
}}

\item{dp}{a \code{numeric} variable, number of decimal places to display.}

\item{p1}{a \code{string} variable, the separator appeared between median and Q1. i.e. \code{<median><p1><iqr><p2>}}

\item{p2}{a \code{string} variable, the separator appeared between Q1 and Q3. i.e. \code{<median><p1><Q1><p2><Q3><p3>}}

\item{p3}{a \code{string} variable, the separator appeared after Q3. i.e. \code{<median><p1><Q1><p2><Q3><p3>}}

\item{median.iqr.pars}{a \code{character} vector which stores all parameter name that need median iqr calculation.}
}
\description{
This function gives a table of listing parameters' median and iqr
}
\examples{
\dontrun{
# d<-dat
# p<-get.stat.par(par,par$Parameters))
# median.iqr.p<-as.vector(p[["median.iqr."]])
# median.iqr.dp<-2
# median.iqr.p1<-"("
# median.iqr.p2<-","
#' # median.iqr.p3<-")"
# median_iqr(d,median.iqr.p,median.iqr.dp,median.iqr.p1,median.iqr.p2,median.iqr.p3)}

}
