% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_to_blank}
\alias{na_to_blank}
\title{Replace NA with Blank}
\usage{
na_to_blank(x)
}
\arguments{
\item{x}{Vector to process}
}
\value{
Character vector with NAs replaced by ""
}
\description{
Converts NA values to empty string for table display.
}
\examples{
na_to_blank(c(1, NA, 3))
}
