% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_to_dash}
\alias{na_to_dash}
\title{Replace NA with Dash}
\usage{
na_to_dash(x)
}
\arguments{
\item{x}{Vector to process}
}
\value{
Character vector with NAs replaced by "-"
}
\description{
Converts NA values to "-" for table display.
}
\examples{
na_to_dash(c(1, NA, 3))
}
