% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{number_to_word}
\alias{number_to_word}
\title{Convert Number to Word}
\usage{
number_to_word(x)
}
\arguments{
\item{x}{Integer vector}
}
\value{
Character vector with numbers as words (for 1-10) or as-is
}
\description{
Converts small integers (1-10) to their word equivalents.
}
\examples{
number_to_word(3)
number_to_word(c(1, 5, 15))
}
