% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova-helpers.R
\name{print_anova}
\alias{print_anova}
\title{Print Formatted ANOVA Table}
\usage{
print_anova(anova_model, caption)
}
\arguments{
\item{anova_model}{An anova data frame}

\item{caption}{Table caption}
}
\value{
A kable object
}
\description{
Formats an ANOVA table for LaTeX output with proper p-value formatting.
}
