% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable-helpers.R
\name{print_xtable}
\alias{print_xtable}
\title{Print Xtable with Auto Format Detection}
\usage{
print_xtable(x, ...)
}
\arguments{
\item{x}{An object that can be coerced to xtable}

\item{...}{Additional arguments passed to print.xtable}
}
\description{
Prints an xtable with automatic format detection. For HTML output,
displays in RStudio viewer if available.
}
\examples{
\dontrun{
print_xtable(head(mtcars))
}
}
