% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pseudo}
\alias{pseudo}
\title{Calculate Pseudo R-squared for Model List}
\usage{
pseudo(model_list)
}
\arguments{
\item{model_list}{A list of fitted GLM models}
}
\value{
Character vector of formatted pseudo R-squared values
}
\description{
Extracts McFadden's pseudo R-squared from a list of GLM models
using \code{pscl::pR2()}.
}
\examples{
\dontrun{
m1 <- glm(am ~ wt, data = mtcars, family = binomial)
m2 <- glm(am ~ wt + hp, data = mtcars, family = binomial)
pseudo(list(m1, m2))
}
}
