% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_attributes.R
\name{remove_attributes}
\alias{remove_attributes}
\title{Remove all attributes}
\usage{
remove_attributes(x)
}
\arguments{
\item{x}{dataframe.}
}
\value{
Dataframe without attributes.
}
\description{
Remove all attributes of all the variables in a dataframe.
}
\author{
Nicolas Mangin
}
