% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round1}
\alias{round1}
\title{Round to One Decimal Place}
\usage{
round1(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector rounded to 1 decimal place
}
\description{
Round to One Decimal Place
}
\examples{
round1(3.456)
}
