% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round2}
\alias{round2}
\title{Round to Two Decimal Places}
\usage{
round2(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector rounded to 2 decimal places
}
\description{
Round to Two Decimal Places
}
\examples{
round2(3.456)
}
