% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer-helpers.R
\docType{data}
\name{star_cut_three}
\alias{star_cut_three}
\title{Three-Star Cutoffs}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
star_cut_three
}
\description{
Traditional three-star significance cutoffs.
}
\keyword{datasets}
