% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer-helpers.R
\docType{data}
\name{star_cut_vector}
\alias{star_cut_vector}
\title{Default Star Cutoffs}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
star_cut_vector
}
\description{
Default significance cutoffs for stargazer tables.
By default, only uses p < 0.05 (single star).
}
\keyword{datasets}
