% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer-helpers.R
\name{star_nrm}
\alias{star_nrm}
\title{Stargazer with Normal Font}
\usage{
star_nrm(..., type = NULL, digits = 2, star.cutoffs = star_cut_vector)
}
\arguments{
\item{...}{Arguments passed to stargazer}

\item{type}{Output format. If NULL (default), auto-detected.}

\item{digits}{Number of digits to display}

\item{star.cutoffs}{Significance cutoffs}
}
\description{
Stargazer wrapper using default (normal) font size.
}
