% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer-helpers.R
\name{star_sm}
\alias{star_sm}
\title{Stargazer with Small Font}
\usage{
star_sm(..., type = NULL, digits = 2, star.cutoffs = star_cut_vector)
}
\arguments{
\item{...}{Arguments passed to stargazer}

\item{type}{Output format. If NULL (default), auto-detected.}

\item{digits}{Number of digits to display}

\item{star.cutoffs}{Significance cutoffs}
}
\description{
Stargazer wrapper using small font size instead of scriptsize.
}
