% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tble_make_unique.R
\name{tble_make_unique}
\alias{tble_make_unique}
\title{Force unique observations}
\usage{
tble_make_unique(x)
}
\arguments{
\item{x}{dataframe. Variables indicating the attributes of what should be a single observation (mapped after nesting by id).}
}
\value{
Dataframe with only one observation.
}
\description{
Remove duplicated entries. If there are several discrepant pieces of information about the observation, take the most frequent unique value, or the mean, or a random value of the duplicates to keep only one observation.
}
\author{
Nicolas Mangin
}
