% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tble_spread_gather.R
\name{tble_spread_gather}
\alias{tble_spread_gather}
\title{Complete missing pairs}
\usage{
tble_spread_gather(x, col, val, fill)
}
\arguments{
\item{x}{tibble or dataframe.}

\item{col}{character. Name of the column which should be srepad.}

\item{val}{character. Name of the column populating the cells.}

\item{fill}{NA or numeric. Value used to fill missing data.}
}
\value{
A tibble or dataframe where missing pairs have beend added.
}
\description{
Spread and gather the table so that missing pairs are added.
}
\author{
Nicolas Mangin
}
