% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texcount.R
\name{texcount}
\alias{texcount}
\title{TeXcount Word Count Functions}
\description{
Functions to count words in LaTeX/R Markdown documents using texcount.
Requires texcount to be installed (included with most TeX distributions).
}
