% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{Compute and format univariate statistics}
\usage{
univariate(
  x,
  include = c("Variable", "Missing", "Count", "Min", "Median", "Mean", "Max", "Range",
    "St.Dev", "Skew", "Kurt")
)
}
\arguments{
\item{x}{Tibble or dataframe. Table with only numeric variables.}

\item{include}{Character vector. Names of the statistics to report: "Count", "Min", "Median", "Mean", "Max", "Range", "St.Dev","Skew" and "Kurt".}
}
\value{
Dataframe. Table of univariate statistics.
}
\description{
Generate a table displaying univariate statistics for all the variables included in x which are numeric.
}
\author{
Nicolas Mangin
}
